deploy for now.

+++ pages/api/posts/[postId].ts
@@ -1,7 +1,7 @@
 import formidable from 'formidable';
 import { NextApiHandler } from 'next';
 import cloudinary from '../../../lib/cloudinary';
-import { readFile } from '../../../lib/utils';
+import { isAdmin, readFile } from '../../../lib/utils';
 import { postValidationSchema, validateSchema } from '../../../lib/validator';
 import Post from '../../../models/Post';
 import { IncomingPost } from '../../../utils/types';
@@ -25,6 +25,9 @@ const handler: NextApiHandler = (req, res) => {
 
 const removePost: NextApiHandler = async (req, res) => {
   try {
+    const admin = await isAdmin(req, res);
+    if (!admin) return res.status(401).json({ error: 'Unauthorized' });
+
     const postId = req.query.postId as string;
     const post = await Post.findByIdAndDelete(postId);
 
@@ -41,6 +44,9 @@ const removePost: NextApiHandler = async (req, res) => {
 };
 
 const updatePost: NextApiHandler = async (req, res) => {
+  const admin = await isAdmin(req, res);
+  if (!admin) return res.status(401).json({ error: 'Unauthorized' });
+
   const postId = req.query.postId as string;
   const post = await Post.findById(postId);
   if (!post) return res.status(404).json({ error: 'Post not found!' });
